web-animations/animation-timeline/animation-timeline-get-animation-players-test.html

<!DOCTYPE html>
<html>
<head>
<title>Animation timeline getAnimationPlayers() method test</title>
<meta name="assert" content="Animation timeline getAnimationPlayers() method returns the set of AnimationPlayer objects associated with this timeline that have associated source content which is current">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();

    assert_not_equals(doc.timeline.getAnimationPlayers(), null,
        'getAnimationPlayers() method shouldn\'t return null');
    assert_equals(doc.timeline.getAnimationPlayers().length, 0,
        'Set of objects returned by getAnimationPlayers() method should be empty');
}, 'Test set of objects returned by getAnimationPlayers() method when there\'s no ' +
    'players associated');


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    assert_not_equals(doc.timeline.getAnimationPlayers(), null,
        'getAnimationPlayers() method shouldn\'t return null');
    assert_equals(doc.timeline.getAnimationPlayers().length, 1,
        'Set of objects returned by getAnimationPlayers() method should contain one element');
}, 'Test set of objects returned by getAnimationPlayers() method when there is one ' +
    'associated player');


test(function() {
    var doc = newHTMLDocument();
    var player1 = newAnimationPlayer(doc);
    var player2 = newAnimationPlayer(doc);

    assert_array_equals(doc.timeline.getAnimationPlayers(), [player1, player2],
        'Set of objects returned by getAnimationPlayers() method should contain two elements');
}, 'Test set of objects returned by getAnimationPlayers() method when there are more ' +
    'than one associated players');
</script>
</body>
</html>
